Delivery query and create mutation created

+++ server/app-data/graphql/typeDefs/DeliveryMethode/index.ts
@@ -3,13 +3,13 @@ import { gql } from 'apollo-server-express';
 const DeliveryMethode = gql`
   input DeliveryInput {
     title: String
-    value: Int
+    value: Float
   }
 
   type Delivery {
     _id: String
     title: String
-    value: Int
+    value: Float
   }
 `;
